validaciones
se agregan validaciones

+++ app/src/main/java/com/moappdev/solutions/tiendaapp/MainActivity.kt
@@ -1,9 +1,13 @@
 package com.moappdev.solutions.tiendaapp
 
+import android.content.DialogInterface
+import android.content.Intent
+import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
+import android.widget.Toast
 import androidx.recyclerview.widget.GridLayoutManager
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.moappdev.solutions.tiendaapp.database.Tienda
 import com.moappdev.solutions.tiendaapp.database.TiendaApplication
 import com.moappdev.solutions.tiendaapp.databinding.ActivityMainBinding
@@ -30,18 +34,17 @@ class MainActivity : AppCompatActivity(), OnClickListener, MainAux {
     }
 
 //--------------Funciones------------------------
-
     private fun goToTiendaFragment(args: Bundle?=null){
         val fragment= TiendaFragment()
         if(args!=null) fragment.arguments=args
 
+        hideBtnFab( )
         supportFragmentManager
             .beginTransaction()
             .add(R.id.contrainerMain, fragment)
             .addToBackStack(null)
             .commit()
     }
-
     private fun setupRecyclerView() {
         mAdapter= TiendaAdapter(mutableListOf(), this)
         mGridLayoutManager= GridLayoutManager(this, 2)
@@ -53,7 +56,6 @@ class MainActivity : AppCompatActivity(), OnClickListener, MainAux {
 
         }
     }
-
     private fun getTiendas(){
         doAsync {
             val tiendas= TiendaApplication.database.tiendaDao().allTiendas()
@@ -79,14 +81,59 @@ class MainActivity : AppCompatActivity(), OnClickListener, MainAux {
         }
     }
     override fun onDelete(tienda: Tienda) {
-        doAsync {
-            TiendaApplication.database.tiendaDao().deleteTiienda(tienda)
-            uiThread {
-                mAdapter.delete(tienda)
+        val opciones=arrayOf("Eliminar", "Llamar", "Ir al sitio web")
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.ad_title )
+            .setItems(opciones, DialogInterface.OnClickListener { dialog, which ->
+                when(which){
+                    opciones.indexOf("Eliminar")-> dialogEliminar(tienda)
+                    opciones.indexOf("Llamar")-> dialogLlamar(tienda.telefono)
+                    2-> dialogWeb(tienda.web)
+                }
+            }).show()
+    }
+    private fun dialogEliminar(tienda: Tienda){
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.ad_title)
+            .setPositiveButton(R.string.ad_eliminar_ok, DialogInterface.OnClickListener { dialog, which ->
+                doAsync {
+                    TiendaApplication.database.tiendaDao().deleteTiienda(tienda)
+                    uiThread {
+                        mAdapter.delete(tienda)
+                    }
+                }
+            })
+            .setNegativeButton(R.string.ad_eliminar_no,null)
+            .show()
+    }
+    private fun dialogLlamar(telefono: String){
+       if(telefono.isEmpty())
+           Toast.makeText(this, "esta tienda no tiene un nro de telefono", Toast.LENGTH_SHORT).show()
+        else {
+           val i = Intent().apply {
+               action = Intent.ACTION_DIAL
+               data = Uri.parse("tel:$telefono")
+           }
+           if (i.resolveActivity(packageManager) != null)
+               startActivity(i)
+           else
+               Toast.makeText(this, getString(R.string.msgResolve), Toast.LENGTH_SHORT).show()
+       }
+    }
+    private fun dialogWeb(url: String){
+        if(url.isEmpty())
+            Toast.makeText(this, "esta tienda no tiene un sitio web", Toast.LENGTH_SHORT).show()
+        else {
+            val i = Intent().apply {
+                action = Intent.ACTION_VIEW
+                data = Uri.parse(url)
             }
+            if(i.resolveActivity(packageManager)!=null)
+                startActivity(i)
+            else
+                Toast.makeText(this, getString(R.string.msgResolve), Toast.LENGTH_SHORT).show()
         }
     }
-
 //--------------MainAux---------------
 
     override fun hideBtnFab(visible: Boolean) {
@@ -95,12 +142,10 @@ class MainActivity : AppCompatActivity(), OnClickListener, MainAux {
         else
             mBinding.btnFab.hide()
     }
-
     override fun addTienda(tienda: Tienda) {
         mAdapter.add(tienda)
     }
-
     override fun updateTienda(tienda: Tienda) {
-
+        mAdapter.update(tienda)
     }
 }
